requirements.txt
ipykernel
pyright
ruff
polars
matplotlib
pyyaml